CV editor section edit
