+++ public/index.html
@@ -14,7 +14,7 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
 
-=======
+
     <!-- fontaswome added link  -->
     <script
       src="https://kit.fontawesome.com/c856552904.js"
